requirements.txt
streamlit==1.31.0
Pillow==10.2.0
numpy==1.26.3
requests==2.31.0
